Dynamically allocate memory for UVC frame
With a 1280x720 buffer for the UVC frame, the kernel frequently runs out
of memory. This patch allocates only as much memory as needed and only
when needed.

In a different approach, I tried to allocate a memblock in
uvc_handle_video_streaming_req_recv and free it in
uvc_handle_video_abort, along with using a mutex in these two functions
as well as in send_frame. However, that did not prevent use after free
from occuring occasionally when switching resolution modes, because
mutexes cannot be locked from interrupts.